requirements.txt
boost==0.1 
#cmake==3.25.0
Click>=6.0

Django==4.1
django-cors-headers
django-ckeditor==6.4.2
django-import-export==2.8.0
DateTime==4.5

numpy==1.23.3
# opencv-python==4.5.3.56
#dlib==19.23.0
#./dlib-19.22.99-cp39-cp39-win_amd64.whl
scipy>=0.17.0
imutils==0.5.3
scikit-learn==1.1.2
tensorflow==2.7.2
keras==2.7.0
Pillow==9.0.1
matplotlib
ytmusicapi==0.24.0
lxml[html_clean]
requests-html==0.10.0
channels==4.0.0
daphne==2.5.0
whitenoise==6.4.0
psycopg2-binary
nest-asyncio
# botasaurus
poe_api_wrapper
websocket-client
#mysqlclient==2.1.1
# uvicorn[standard]

pandas>=1.5.0
pydantic>=1.10.2
Faker>=15.1.1
lxml>=4.9.1
prettytable>=3.4.1
typer[all]>=0.6.1
update>=0.0.1
aiohttp>=3.8.3
aiohttp-proxy>=0.1.2
html5lib>=1.1
requests>=2.28.1
bs4>=0.0.1
strenum>=0.4.10
protobuf==3.19.6
lxml_html_clean